Moved gamedetails
